Added an "other" constructor to enum ReportType
I was receiving the following compiler error:

src/massive/munit/command/ReportCommand.hx:168: lines 168-173 : This pattern is unused

because the switch statement had 2 options ('teamcity' and default), but the enum
only had the one option.  I added the 'other' option to make the error go away, but leave
the switch code in place for future ReportTypes.